WORKING-IDE-project + Altium schemes
Project in Cube IDE IS RUNNING :8ball:  and printing string to virtual terminal.
Use it as checkpoint, don't change anthng with pins and interfaces unless absolutely necessary. But if you want - save BACKUPs on your own PC and dont make commits if you are not shure that project is still working.